requirements.txt
fastapi
uvicorn
pydantic
passlib[bcrypt]
PyJWT
python-multipart